+++ spec/models/concerns/federails/data_entity_spec.rb
@@ -62,5 +62,16 @@ module Federails
         end
       end
     end
+
+    describe 'Inbox hook' do
+      it 'successfully defines an Inbox hook for incoming Create activities' do
+        allow(Fixtures::Classes::FakeDataModel).to receive(:handle_incoming_fediverse_data)
+
+        Fediverse::Inbox.dispatch_request 'type' => 'Create', 'object' => { 'type' => 'TestThing' }
+        Fediverse::Inbox.dispatch_request 'type' => 'Create', 'object' => { 'type' => 'OtherThing' }
+
+        expect(Fixtures::Classes::FakeDataModel).to have_received(:handle_incoming_fediverse_data).once
+      end
+    end
   end
 end